requirements.txt
kafka-python==2.0.2
toolz==1.0.0
